Move package from cmd/perunode -> node
Move the package implementing node API to node, so that the package
(cmd/perunode) can be used for the executable package implementing
perunnode app.

Relates to #98.
