+++ testdata/node/invalid.yaml
