+++ testdata/node/valid.yaml
